requirements.txt
Django==4.2.2
django-parler==2.3
djangorestframework==3.14.0
coreapi==2.3.3